Rerevise numbers for year-to-year antigenic drift.
I did not like the holes in the dataset that previously existed because
H1 and B had years with no virus isolations.  Here, I've sliced the tree
at each year and taken the mean location on dimension 1 of these slices.
This also allowed me to take the surveillance data back to 1998.